Move LoginController to separate class
